backend/requirements.txt
openai>=1.0.0,<2.0.0
alembic==1.14.0
bcrypt==4.2.1
fastapi==0.115.5
PyMuPDF==1.24.13
PyJWT==2.10.1
psycopg[binary]==3.2.3
python-docx==1.1.2
python-multipart==0.0.17
scikit-learn==1.5.2
spacy==3.8.12
SQLAlchemy==2.0.36
uvicorn==0.32.0
